+++ GameStore-Backend/src/main/java/ca/mcgill/ecse321group1/gamestore/repository/ReviewRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import ca.mcgill.ecse321group1.gamestore.model.Review;
 import java.util.Optional;
 
-public interface ReviewRepository extends CrudRepository<Review, Long> {
-	Review findReviewById(Long id);
+public interface ReviewRepository extends CrudRepository<Review, Integer> {
+	Review findReviewById(int id);
 }
