+++ GameStore-Backend/src/main/java/ca/mcgill/ecse321group1/gamestore/repository/VideoGameRepository.java
@@ -4,8 +4,8 @@ import org.springframework.data.repository.CrudRepository;
 
 import ca.mcgill.ecse321group1.gamestore.model.VideoGame;
 
-public interface VideoGameRepository extends CrudRepository<VideoGame, Long> {
+public interface VideoGameRepository extends CrudRepository<VideoGame, Integer> {
 
-	VideoGame findVideoGameById(Long id);
+	VideoGame findVideoGameById(int id);
 
 }
